+++ apps/backend/src/modules/user-routes/user-route.service.ts
@@ -9,6 +9,7 @@ import {
 } from "./libs/enums/enum.js";
 import { UserRouteError } from "./libs/exceptions/exceptions.js";
 import {
+	type UserRouteFilter,
 	type UserRouteResponseDto,
 	type UserRouteStatusType,
 } from "./libs/types/type.js";
@@ -34,6 +35,10 @@ class UserRouteService implements Service {
 	}): Promise<UserRouteResponseDto> {
 		const { routeId, userId } = payload;
 
+		await this.ensureUserIsNotOnActiveRoute(userId);
+
+		await this.ensureIsNotDuplicateRoute(routeId, userId);
+
 		const { geometry } = await this.routeService.findById(routeId);
 
 		const createdData = UserRouteEntity.initializeNew({
@@ -43,9 +48,6 @@ class UserRouteService implements Service {
 			status: UserRouteStatus.NOT_STARTED,
 			userId,
 		});
-
-		await this.ensureIsNotDuplicateRoute(routeId, userId);
-
 		const createdRoute = await this.userRouteRepository.create(createdData);
 
 		return createdRoute.toObject();
@@ -77,7 +79,10 @@ class UserRouteService implements Service {
 	}): Promise<UserRouteResponseDto> {
 		const { actualGeometry, routeId, userId } = payload;
 
-		const userRoute = await this.getByRouteIdAndUserId(routeId, userId);
+		const userRoute = await this.getRouteByFilter({
+			routeId,
+			userId,
+		});
 
 		this.ensureUserIsOwner(userRoute.userId, userId);
 
@@ -112,13 +117,31 @@ class UserRouteService implements Service {
 		return userRoutes.map((item) => item.toObject());
 	}
 
+	public async getRouteByFilter(
+		filters: UserRouteFilter,
+	): Promise<UserRouteResponseDto> {
+		const [userRoute] = await this.userRouteRepository.findByFilter(filters);
+
+		if (!userRoute) {
+			throw new UserRouteError({
+				message: UserRouteExeptionMessage.USER_ROUTE_NOT_FOUND,
+				status: HTTPCode.NOT_FOUND,
+			});
+		}
+
+		return userRoute.toObject();
+	}
+
 	public async start(payload: {
 		routeId: number;
 		userId: number;
 	}): Promise<UserRouteResponseDto> {
 		const { routeId, userId } = payload;
 
-		const userRoute = await this.getByRouteIdAndUserId(routeId, userId);
+		const userRoute = await this.getRouteByFilter({
+			routeId,
+			userId,
+		});
 
 		this.ensureUserIsOwner(userRoute.userId, userId);
 
@@ -154,6 +177,7 @@ class UserRouteService implements Service {
 	): Promise<void> {
 		const userRoutes = await this.userRouteRepository.findByFilter({
 			routeId,
+			status: UserRouteStatus.NOT_STARTED,
 			userId,
 		});
 
@@ -193,25 +217,6 @@ class UserRouteService implements Service {
 				status: HTTPCode.FORBIDDEN,
 			});
 		}
-	}
-
-	private async getByRouteIdAndUserId(
-		routeId: number,
-		userId: number,
-	): Promise<UserRouteResponseDto> {
-		const [userRoute] = await this.userRouteRepository.findByFilter({
-			routeId,
-			userId,
-		});
-
-		if (!userRoute) {
-			throw new UserRouteError({
-				message: UserRouteExeptionMessage.USER_ROUTE_NOT_FOUND,
-				status: HTTPCode.NOT_FOUND,
-			});
-		}
-
-		return userRoute.toObject();
 	}
 }