+++ apps/frontend/src/libs/modules/storage/libs/enums/storage-key.enum.ts
@@ -1,4 +1,5 @@
 const StorageKey = {
+	ACTUAL_PATH_COORDINATES: "actual-path-coordinates",
 	CREATE_ROUTE_FORM_DATA: "create-route-form-data",
 	TOKEN: "token",
 } as const;