+++ apps/frontend/src/libs/modules/store/store.module.ts
@@ -33,7 +33,11 @@ import {
 	routesApi,
 	routesReducer,
 } from "~/modules/routes/routes.js";
-import { userRoutesApi } from "~/modules/user-routes/user-routes.js";
+import {
+	userRouteApi,
+	userRouteDetailsReducer,
+	userRoutesReducer,
+} from "~/modules/user-routes/user-routes.js";
 import { userApi, reducer as usersReducer } from "~/modules/users/users.js";
 
 import { type ExtraArguments, type RootReducer } from "./libs/types/types.js";
@@ -59,7 +63,7 @@ class Store {
 			storage,
 			toastNotifier,
 			userApi,
-			userRoutesApi,
+			userRouteApi,
 		};
 	}
 
@@ -85,6 +89,8 @@ class Store {
 				pointsOfInterest: pointsOfInterestReducer,
 				routeDetails: routeDetailsReducer,
 				routes: routesReducer,
+				userRouteDetails: userRouteDetailsReducer,
+				userRoutes: userRoutesReducer,
 				users: usersReducer,
 			},
 		});