+++ apps/frontend/src/modules/user-routes/libs/constants/constants.ts
@@ -1,0 +1,1 @@
+export { LOCATION_TRACKING_CONFIG } from "./location-tracking-config.constant.js";