+++ apps/frontend/src/modules/user-routes/libs/enums/enums.ts
@@ -1,1 +1,2 @@
-export { UserRouteApiPath } from "@smartscapes/shared";
+export { TrackingExeptionMessage } from "./tracking-exeption-message.enum.js";
+export { UserRouteApiPath, UserRouteStatus } from "@smartscapes/shared";