+++ apps/frontend/CHANGELOG.md
@@ -1,5 +1,24 @@
 # Changelog
 
+## [1.53.0](https://github.com/BinaryStudioAcademy/bsa-2025-smartscapes/compare/@smartscapes/frontend-v1.52.0...@smartscapes/frontend-v1.53.0) (2025-08-27)
+
+
+### Features
+
+* **frontend:** user route navigate ss-501 ([#558](https://github.com/BinaryStudioAcademy/bsa-2025-smartscapes/issues/558)) ([8d56411](https://github.com/BinaryStudioAcademy/bsa-2025-smartscapes/commit/8d56411cea39997ded2c8a0545397df9d28638ff))
+
+
+### Bug Fixes
+
+* **frontend:** swap poi coordinates to correctly display location ss-571 ([#572](https://github.com/BinaryStudioAcademy/bsa-2025-smartscapes/issues/572)) ([58e4561](https://github.com/BinaryStudioAcademy/bsa-2025-smartscapes/commit/58e4561b88cb2e4128abc1f716895d9f8b13f719))
+
+
+### Dependencies
+
+* The following workspace dependencies were updated
+  * dependencies
+    * @smartscapes/shared bumped from 1.38.0 to 1.39.0
+
 ## [1.52.0](https://github.com/BinaryStudioAcademy/bsa-2025-smartscapes/compare/@smartscapes/frontend-v1.51.0...@smartscapes/frontend-v1.52.0) (2025-08-27)